rename(calculator package): 각 클래스를 view 및 controller 패키지로 이동
