+++ src/main/java/calculator/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package exception;
+package calculator.exception;
 
 
 public enum ErrorMessage {
